+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - README.md tracking deployed Versions
 - tests to fit new protocol developments
 - utils to use governance contract more easily
+- Fund function to retrieve order ID by exchange/asset pair
 
 ### Changed